python/requirements.txt
# requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
pdf2image==1.16.3
pytesseract==0.3.10
opencv-python==4.8.1.78
numpy==1.24.3
pillow==10.1.0
pypdf2==3.0.1
pymupdf==1.23.8
transformers==4.35.0
torch==2.1.0
easyocr==1.7.1
layoutparser==0.3.4
detectron2 @ git+https://github.com/facebookresearch/detectron2.git
celery==5.3.4
redis==5.0.1
boto3==1.34.0
python-dotenv==1.0.0  